Respawn system implemented. Potential design flaw, but works functionally.
Potential Design Flaw: I have the respawning logic incorporated into AC_Character, which is a component. It might be better practice to have respawning handled by the level, which can allow for the player to destroy and repossess a new character and manage lives as an array rather than individually. However, the current system should suffice, as there is not a large difference between the two systems besides level of control, since the level calls for the player to call its respawn function anyways.